chore: normalize `systemId` and `system_id` to `system`

+++ packages/@kblocks/control/src/delete.ts
@@ -3,9 +3,14 @@ import { type ErrorEvent, emitEvent, formatBlockType, parseBlockUri } from "./ap
 import { Context } from "./context";
 
 export async function deleteObject(client: k8s.CustomObjectsApi, ctx: Context, objUri: string) {
-  const { group, version, plural, system: systemId } = ctx;
+  const { group, version, plural, system } = ctx;
+
   const blockUri = parseBlockUri(objUri);
   const blockType = formatBlockType(blockUri);
+
+  if (system !== blockUri.system) {
+    throw new Error(`Control message sent to wrong system. My system is ${system} but the message is for ${blockUri.system}`);
+  }
 
   try {
     await client.deleteNamespacedCustomObject(group, version, blockUri.namespace, plural, blockUri.name);
